test files containg threshold comparison, etc. are added
